Update Thought hash
micro save. working throughout the domain specific relationships that are directly related and indirectly related. This is a loose brain dump of the subject matter and research of the particular overarching architectural archetypal themes that will serve as the background back of house (like a commercial kitchen) for jump start. at some point there will be a high level supersets an automation bus logic for this logic, then another superset with a custom built neronet.